Kerbrute + ZSH path fixes

+++ user/local/bin/update-kerbrute.sh
@@ -15,17 +15,21 @@ REMOTE_VERSION="$(curl -L -s https://api.github.com/repos/ropnop/kerbrute/releas
 # Install/update if there is a new version.
 #
 if [[ "$LOCAL_VERSION" != "$REMOTE_VERSION" ]]; then
-	BUILD_DIR="$(mktemp -d)"
-	(
-		cd "$BUILD_DIR"
-		curl -L -O https://github.com/ropnop/kerbrute/releases/download/v${REMOTE_VERSION}/kerbrute_linux_amd64
-		mkdir -p $HOME/.local/bin
-		mv kerbrute_linux_amd64 $HOME/.local/bin/kerbrute
-		chmod +x $HOME/.local/bin/kerbrute
-		mkdir -p $HOME/.cache/versions
-		echo "$REMOTE_VERSION" > $HOME/.cache/versions/kerbrute
-	)
-	rm -rf "$BUILD_DIR"
+	if [[ "$(uname -m)" == "x86_64" ]]; then
+		BUILD_DIR="$(mktemp -d)"
+		(
+			cd "$BUILD_DIR"
+			curl -L -O https://github.com/ropnop/kerbrute/releases/download/v${REMOTE_VERSION}/kerbrute_linux_amd64
+			mkdir -p $HOME/.local/bin
+			mv kerbrute_linux_amd64 $HOME/.local/bin/kerbrute
+			chmod +x $HOME/.local/bin/kerbrute
+		)
+		rm -rf "$BUILD_DIR"
+	else
+		go install github.com/ropnop/kerbrute@v${REMOTE_VERSION}
+	fi	
+	mkdir -p $HOME/.cache/versions
+	echo "$REMOTE_VERSION" > $HOME/.cache/versions/kerbrute
 else
 	echo "Kerbrute is already at v${REMOTE_VERSION}"
 	touch $HOME/.cache/versions/kerbrute
